リファクタリング: 型定義を別ファイルに切り出し
